producer/requirements.txt
binance-futures-connector
kafka-python
